new-microservice/requirements.txt
flask
psycopg2-binary
requests
flake8